error 1

+++ flutter_app/lib/features/auth/presentation/provider_register_screen.dart
@@ -805,8 +805,7 @@ class _ProviderRegisterScreenState extends State<ProviderRegisterScreen> {
       return 'Service';
     }
 
-    final safeIndex =
-        _serviceDetailIndex.clamp(0, _selectedServices.length - 1) as int;
+    final safeIndex = _serviceDetailIndex.clamp(0, _selectedServices.length - 1);
     return _selectedServices[safeIndex];
   }
